Attempt re-implementation in Go
The rationale is this:
* `px` on OS X needs root privileges to be able to list memory usage, CPU usage
and command lines for other processes
* Having to run `px` using `sudo` is not optimal
* We would like to `setuid` the `px` binary to `root`
* The `setuid` bit on a Python script (as produced by Pants) has no effect

As an alternative, Go produces statically linked binaries that we can `setuid`
just fine. Also, Go happens to come with [a port of the same process listing
library we're using in Python anyway](https://github.com/shirou/gopsutil).

So let's try out Go and see where this takes us.

A note about at least the initial Go commits: I haven't written any Go prior to
this, so the first steps will likely be small.

This first commit is just a copy of the example from the Go Psutil package.