Make top tests less prone to intermittent failures
Before this change we failed unless at least one line in the top output
had been cut.

With this change in place, we check the header line length explicitly,
then verify that no process information line is too long.

We also make the screen narrower during the test to increase the
likelihood of the top display actually having to shorten some line(s).