Suggest "sudo lsof" rather than just "lsof"
When invoking px on a single PID, suggest using "sudo lsof" rather than
just "lsof" to get the full list of file descriptors for a process.

I sometimes copy / paste these suggestions, and a lot of times when I do
that I have to add "sudo" manually to get it working.