Fix post-release duplicate wheels problem
During release, wheels with proper version numbers are created.

During normal build, wheels with version number 0.0.0 are created.

Having the 1.0.36 wheels around during ordinary builds messes up pex
creation, because of a "*.whl" expression.

With this change in place, dist dirs are removed before wheeling and
pexing, and the problem is gone.