ptop: Sort first frame by CPU percentage
Before this change, the first frame was sorted by CPU time used, just
like all the other frames.

However, for the first frame, no processes have used any CPU, so they
were all at 0, and the sorting was really a no-op for this frame.

With this change in place, we sort the first frame by CPU percentage, as
an approximation of upcoming CPU usage numbers.

This makes the first frame look more than frame 2, which makes the
transition between them less jarring to look at.