Rethink future process checks
We used to check for this at runtime, but that would randomly fail in
CI. Since I don't understand why, and therefore can't fix it, let's just
not check for this in CI.

Also, remove the runtime check since that may or may not fail the same
way without me having any idea about how to debug it.

If it fails locally during testing, we'll still be notified.